Fixed *string_copy_to_integer() converting invalid decimal strings
During decimal conversion in the check for if a character is an invalid
digit, the check would always evaluate to false. This allowed
*string_copy_to_integer() to always return a number on decimal
conversion and never fail regardless of sanity of input.

This was discovered during use of libfvde where it was trying to read a
hex plist value as a decimal and returning an incorrect number. Until
libfplist is updated to handle hex integer values libfvde will likely be
unable to deal with disks that it previously could.